+++ src/interfaces/SilverboxFile.ts
