+++ src/interfaces/SilverboxImageParams.ts
