+++ src/lib/Silverbox.spec.ts
@@ -2,8 +2,8 @@ import test from 'ava'
 import Silverbox from './Silverbox'
 
 const CREDENTIALS = {
-  host: 'https://cdn.heseya.com',
-  client: 'heseya',
+  host: 'https://cdn.example.com',
+  client: 'example',
   key: 'key',
 }
 
@@ -15,14 +15,14 @@ test('Silverbox creates succesfully', (t) => {
 test('Silverbox change host', (t) => {
   const silverbox = new Silverbox(CREDENTIALS)
   t.is(silverbox.getHost(), CREDENTIALS.host)
-  silverbox.setHost('https://heseya.com')
+  silverbox.setHost('https://example.com')
   t.not(silverbox.getHost(), CREDENTIALS.host)
 })
 
 test('Silverbox change client', (t) => {
   const silverbox = new Silverbox(CREDENTIALS)
   t.is(silverbox.getClient(), CREDENTIALS.client)
-  silverbox.setClient('https://heseya.com')
+  silverbox.setClient('https://example.com')
   t.not(silverbox.getClient(), CREDENTIALS.client)
 })
 
@@ -37,14 +37,14 @@ test('Silverbox clone instance', (t) => {
 test('Silverbox returns full file path', (t) => {
   const silverbox = new Silverbox(CREDENTIALS)
   const path = silverbox.get('image.png')
-  t.is(path, 'https://cdn.heseya.com/heseya/image.png')
+  t.is(path, 'https://cdn.example.com/example/image.png')
 })
 
 test('Silverbox get file as other client', (t) => {
   const silverbox = new Silverbox(CREDENTIALS)
   const path = silverbox.as('test').get('image.jpg')
 
-  t.is(path, 'https://cdn.heseya.com/test/image.jpg')
+  t.is(path, 'https://cdn.example.com/test/image.jpg')
 })
 
 test('Silverbox no host provided', async (t) => {
